update THIRDPARTY.yml
spdx and its smallvec dependency are new

Now that the license expression is taken as specified without being normalized a bunch of them changed accordingly.

In a few cases this changed the order of the licenses.

The dependencies deunicode, winapi-i686-pc-windows-gnu, and winapi-x86_64-pc-windows-gnu had "NOT FOUND" license text, the corresponding hunks have been ignored to keep the manually specified license text intact